renaming async_manger => spawner_manager
